backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
pydantic-settings==2.6.1
pymongo==4.8.0
python-dotenv==1.0.1
